Rename Find Address and Geocode Online samples
