+++ src/main/java/com/esri/samples/search/find_address/README.md
@@ -1,43 +1,37 @@
 <h1>Find Address</h1>
 
-<p>Demonstrates how to perform geocoding and reverse geocoding with offline data.</p>
+<p>Demonstrates how to geocode an address query and display its location on the ArcGISMap.</p>
 
 <p><img src="FindAddress.png"/></p>
 
 <h2>How to use the sample</h2>
 
-<p>For simplicity, the sample comes loaded with a set of addresses. You can select any address to perform geocoding and show it's location on map. To perform reverse geocoding in real-time, click on any location on the ArcGISMap to provide it's address.</p>
+<p>For simplicity, the sample comes loaded with a set of addresses. You can select an address to perform online geocoding and show the matching results on the ArcGISMap. </p>
 
 <h2>How it works</h2>
 
-<p>To perform geocoding with offline resources:</p>
+<p>To get a geocode from a query and display its location on the <code>ArcGISMap</code>:</p>
 
 <ol>
-    <li>Create an <code>ArcGISMap</code>'s using a <code>Basemap</code>.
-        <ul><li>basemap is created using a <code>TileCache</code>, which represent our offline resource</li></ul></li>
-    <li>Add the map to the <code>MapView</code>, <code>MapView.setMap()</code>. </li>
-    <li>Create a <code>LocatorTask</code> using a URI to the offline locator file and define the <code>ReverseGeocodeParameters</code>/<code>GeocodeParameters</code> for  the LocatorTask.</li>
-    <li>To geocode an address, set the geocode parameters and use <code>LocatorTask.geocodeAsync(geocodeParameters)</code>.</li>
-    <li>To reverse geocode a location, get the <code>Point</code> location on the map view and use <code>LocatorTask.reverseGeocodeAsync(Point)</code>.</li>
-    <li>Show the retrieved results by creating a <code>PictureMarkerSymbol</code> with attributes from the result and add that symbol to a <code>Graphic</code>  in the <code>GraphicsOverlay</code>.</li>
+  <li>Create the ArcGIS map's with <code>Basemap</code>.
+    <ul><li>basemap is created using a <code>TileCache</code> to represent an offline resource </li></ul></li>
+  <li>Create a <code>LocatorTask</code> using a URL.</li>
+  <li>Set the <code>GeocodeParameters</code> for the locator task and specify the geocodes' attributes.</li>
+  <li>Get the matching results from the <code>GeocodeResult</code> using <code>LocatorTask.geocodeAsync(query, geocodeParameters)</code>.</li>
+  <li>Lastly, to show the results using a <code>PictureMarkerSymbol</code> with attributes and add the symbol to a <code>Graphic</code> in the  <code>GraphicsOverlay</code>.</li>
 </ol>
 
 <h2>Features</h2>
 
 <ul>
-    <li>ArcGISMap</li>
-    <li>ArcGISTiledLayer</li>
-    <li>Callout</li>
-    <li>MapView</li>
-    <li>LocatorTask </li>
-    <li>GeocodeParameters</li>
-    <li>GeocodeResult</li>
-    <li>Graphic</li>
-    <li>GraphicsOverlay</li>
-    <li>Point</li>
-    <li>PictureMarkerSymbol</li>
-    <li>ReverseGeocodeParameters</li>
-    <li>TileCache</li>
+  <li>ArcGISMap</li>
+  <li>GeocodeParameters</li>
+  <li>GeocodeResult</li>
+  <li>GraphicsOverlay</li>
+  <li>LocatorTask</li>
+  <li>MapView</li>
+  <li>ReverseGeocodeParameters</li>
+  <li>TileCache</li>
 </ul>
